Generation service, controller and tests added

+++ energy_generation_mix_backend/src/main/java/energymix/client/dto/GenerationResponseDto.java
@@ -12,7 +12,7 @@ public record GenerationResponseDto(List<IntervalData> data) {
 
     public record GenerationMix(
             String fuel,
-            float perc
+            double perc
     ) {}
 
 }
